refactor(fs): extract error-builders to fs/errors.rs (under 2000-cap)
